style: apply Evangelion Title Card design system v4.0
- Add EVA overlays (grain, flicker) to all pages
- Update typography with compressed serif titles (scaleY 0.7, scaleX 0.85)
- Apply status colors (#00ff00, #ffb800, #ff0000) with glow effects
- Add EvaGrainOverlay, EvaFlickerOverlay, EvaTitle, EvaLabel components
- Update all UI components to use EVA color palette
- Add design exploration pages (anti-design, blueprint, claymation, evangelion, newspaper, posthog)
- Update globals.css with EVA theme variables and utilities

+++ packages/rs-shared/src/protocol.rs
@@ -18,7 +18,7 @@ pub struct TicketPayload {
     /// User's subscription plan
     pub plan: String,
     /// Maximum number of concurrent tunnels allowed
-    pub max_tunnels: u8,
+    pub max_tunnels: u32,
     /// Reserved subdomain (None = random assignment)
     #[serde(skip_serializing_if = "Option::is_none")]
     pub subdomain: Option<String>,
